renamed NPIValidationUtil to NPIUtil and moved generate NPI to NPIUtil
